4.1 - Elite Ultra rebalance/fix
As a balancing measure, Elite Ultras no longer have an armor ability. Because of this, it's no longer needed to switch their armor ability when they change berserk states, which inadvertently fixes an issue where the ability will fail to be picked up by them, causing it to spawn every tick and eventually overload the map.